chore: transform workflows
