+++ ui/app/src/components/NewWorkflowNext/Whiteboard.tsx
@@ -16,8 +16,8 @@
  */
 import Menu from '@/mui-extends/Menu'
 import Paper from '@/mui-extends/Paper'
-import { useStoreDispatch, useStoreSelector } from '@/store'
 import { useComponentActions } from '@/zustand/component'
+import { useWorkflowActions, useWorkflowStore } from '@/zustand/workflow'
 import ContentCopyIcon from '@mui/icons-material/ContentCopy'
 import DeleteIcon from '@mui/icons-material/Delete'
 import { Drawer, IconButton, ListItemIcon, ListItemText, MenuItem } from '@mui/material'
@@ -29,8 +29,6 @@ import { MarkerType, Node, ReactFlowInstance, XYPosition } from 'react-flow-rend
 import ReactFlow, { Background, Controls, MiniMap, addEdge, useEdgesState, useNodesState } from 'react-flow-renderer'
 import { useIntl } from 'react-intl'
 import { v4 as uuidv4 } from 'uuid'
-
-import { importNodes, removeWorkflowNode, updateWorkflowNode } from '@/slices/workflows'
 
 import AutoForm, { Belong } from '@/components/AutoForm'
 import i18n, { T } from '@/components/T'
@@ -174,8 +172,8 @@ export default function Whiteboard({ flowRef }: WhiteboardProps) {
   const nodeTypes = useMemo(() => ({ flowNode: FlowNode, groupNode: GroupNode }), [])
   const edgeTypes = useMemo(() => ({ adjustableEdge: AdjustableEdge }), [])
 
-  const store = useStoreSelector((state) => state.workflows)
-  const dispatch = useStoreDispatch()
+  const nodesInStore = useWorkflowStore((state) => state.nodes)
+  const { importNodes, removeWorkflowNode, updateWorkflowNode } = useWorkflowActions()
 
   const [openDrawer, setOpenDrawer] = useState(false)
   const [identifier, setIdentifier] = useState<Identifier | null>(null)
@@ -257,7 +255,7 @@ export default function Whiteboard({ flowRef }: WhiteboardProps) {
       return
     }
 
-    const workflowNode = store.nodes[data.name]
+    const workflowNode = nodesInStore[data.name]
 
     formInitialValues.current = workflowNode
 
@@ -302,7 +300,7 @@ export default function Whiteboard({ flowRef }: WhiteboardProps) {
         return node
       }),
     )
-    dispatch(updateWorkflowNode(rest))
+    updateWorkflowNode(rest)
 
     cleanup()
   }
@@ -363,7 +361,7 @@ export default function Whiteboard({ flowRef }: WhiteboardProps) {
       // Remove templates if they are not used by other nodes.
       templates.forEach((template) => {
         if (!restNodes.some((n) => n.data.name === template)) {
-          dispatch(removeWorkflowNode(template))
+          removeWorkflowNode(template)
         }
       })
 
@@ -399,7 +397,7 @@ export default function Whiteboard({ flowRef }: WhiteboardProps) {
   const importWorkflow = (workflow: string) => {
     const { store, nodes, edges } = workflowToFlow(workflow)
 
-    dispatch(importNodes(store))
+    importNodes(store)
     setNodes(
       nodes.map((node) => ({
         ...node,
